refactor(webui): isolate desktop auto scroll
